--- src/artikkelit/generate/generate084Fields.spec.js
@@ -1,34 +1,0 @@
-import {READERS} from '@natlibfi/fixura';
-import {expect} from 'chai';
-import generateTests from '@natlibfi/fixugen';
-import {generatef084} from './generate0xxFields.js';
-
-generateTests({
-  callback,
-  path: [__dirname, '..', '..', '..', 'test-fixtures', 'generatef084'],
-  useMetadataFile: true,
-  recurse: false,
-  fixura: {
-    reader: READERS.JSON
-  }
-});
-
-async function callback({getFixture, expectToFail = false}) {
-  try {
-    const input = getFixture('input.json');
-    const expectedResults = getFixture('output.json');
-    const result = await generatef084(input.otherRatings);
-
-    expect(result).to.deep.equal(expectedResults);
-    expect(expectToFail, 'This is expected to succes').to.equal(false);
-
-  } catch (error) {
-
-    if (!expectToFail) {
-      throw error;
-    }
-
-    expect(expectToFail, 'This is expected to fail').to.equal(true);
-  }
-
-}
