+++ test-fixtures/generatef080/01/metadata.json
@@ -1,6 +1,7 @@
 {
   "description": "generatef080; field 080",
-  "disabled": false,
+  "skip": false,
+  "only": false,
   "succesRecords": 2,
   "failedRecords": 0,
   "expectedError": false,
