+++ test-fixtures/generatef084/02/metadata.json
@@ -1,5 +1,7 @@
 {
   "description": "This test expected to fail: 422",
+  "skip": false,
+  "only": false,
   "expectToFail": true,
   "expectedFailStatus": 422
 }
