+++ test-fixtures/generatef084/03/metadata.json
@@ -1,5 +1,7 @@
 {
   "description": "This test expected to fail: 400",
+  "skip": false,
+  "only": false,
   "expectToFail": true,
   "expectedFailStatus": 400
 }
